Fix for turris-diagnostics version 13.0.0
Turris Diagnostics added new listing output for modules. Format is that
module names are just simply listed on single line and description of
them is on subsequent lines indented by one space. There is no other
output in this mode.

One more change is that number for diagnostics are now internal and are
not required to be specified. Diagnostics use them to display modules in
given order but they are not reported. This means that order in
resulting dictionary is the order given by those numbers.

Last but most important change is that 'help' argument was removed. That
forces this change to be deployed alongside with turris-diagnostics
version 13.0.0 to prevent breakage.